[RFC] Establish OpenXLA Code of Conduct (#94)
The OpenXLA project currently has no written down code of conduct. We
have discussed this among core maintainers and agreed on corner stones
of a code of conduct. I have assembled a proposal based on commonly used
write ups that covers what we consider crucial.

Please comment or propose changes if there is an area we have not
covered or that needs clarifying. Note that we will do a follow up RFC
on enforcement.